Refactor host file location and add hosts array script
The location of the hosts.txt file has been changed from a specific package directory to the root directory. This change is reflected in all scripts that reference this file. Additionally, a new Node.js script has been added to generate an array of hosts from the hosts.txt file. The Makefile has also been updated with a new target for running this script.