Init stuff (#1)
* Setting up repo stuff

* Updated Makefile with install dependencies

The Makefile has been updated to include new 'install' dependencies. The hosts now depend on the 'install' target, which ensures that yarn and corepack are properly installed before running Pulumi preview. This change improves the build process by ensuring necessary tools are available before execution.

* Update root yarn lock

* Added renovate configuration

Introduced a new configuration file for Renovate bot. This includes enabling automerge, defining custom managers with regex patterns for dependency name templates and match strings, and setting up package rules.

* Refactor host file location and add hosts array script

The location of the hosts.txt file has been changed from a specific package directory to the root directory. This change is reflected in all scripts that reference this file. Additionally, a new Node.js script has been added to generate an array of hosts from the hosts.txt file. The Makefile has also been updated with a new target for running this script.

* Add a workflow

* Building out the components package

* Updated Makefile dependencies and build process

The Makefile has been updated to include a new 'components' target, which is now a dependency for the 'preview' and '$(HOSTS)' targets. This change ensures that components are built before these tasks are run.

Additionally, in the components package's Makefile, we've added commands to create a '.make' directory if it doesn't exist and to export the root directory of the git repository. A new 'build' target has also been introduced which depends on the 'install' target.

* Updated Node version

The Node version has been updated to 22.4.1 in the .nvmrc file for better compatibility and performance improvements.

* Format

* Added recommended extensions and settings for VSCode

Introduced a new `.vscode/extensions.json` file with a list of recommended extensions for the project. Also added a `.vscode/settings.json` file to set `dprint.dprint` as the default formatter and enable its experimental LSP feature.

* Added new dependency and updated yarn.lock

A new package '@pulumi/tls' has been added to the project dependencies in package.json. The version of this package is set to '^5.0.4'. Correspondingly, the yarn.lock file has also been updated to reflect this change.

* Refactor code for better organization and readability

- Moved type definitions to a new 'types.ts' file
- Extracted the readFiles function into a utility module as readVersions
- Replaced direct usage of zod with imported types from 'types.ts'
- Removed redundant code in config.ts by using the extracted utility function and imported types.

* Fixed typo in workspace path and added new dependency

Corrected a typographical error in the workspace path from 'componenets' to 'components'. Also, a new dependency 'host-components' version 0.0.1 has been added to the hosts package.

Update lockfile

* Copy over configuration

* Added version configuration to Pulumi.yaml

The Pulumi.yaml file has been updated with a new 'versions' section. This includes specific versions for containerd, k8s, ksca, talos, runc, crictl, cniPlugins and krel. These changes ensure that the project uses consistent software versions across different environments.

* Add fp-ts

* Moving things around

* Added configuration handling for hosts

In this update, we've introduced a new way to handle configurations in the host packages. We now import and use the Config object from '@pulumi/pulumi' to manage our settings. This includes requiring objects like 'ethernets' and 'hostname'.

Additionally, we've started laying out some groundwork for future features related to networking components such as netplan, bond, and vlan. These are currently commented out but will be activated in upcoming updates.

This change should improve how we manage configurations across different parts of our application.

* Updated command in workflow file

Modified the command used to generate hosts array in the GitHub Actions workflow. The previous version was not correctly outputting the results, so it has been replaced with an echo statement for proper execution and logging.

* Updated concurrency group in workflow

The concurrency group in the GitHub Actions workflow has been updated to include the matrix host. This change ensures that jobs running on different hosts do not cancel each other, providing a more efficient and reliable CI/CD process.

* Updated workflow concurrency and linting conditions

The concurrency group in the GitHub Actions workflow has been simplified to only consider the matrix host. The condition for running the lint step has also been modified, effectively disabling it for now.

* Set work dir and remove pulumi install step